add Classifier to SafeErrWrapperBuilder
